+++ tests/cli.rs
@@ -139,6 +139,42 @@ fn command_merge() -> Result<(), Box<dyn std::error::Error>> {
 }
 
 #[test]
+fn command_unitigs() -> Result<(), Box<dyn std::error::Error>> {
+    let mut cmd = Command::cargo_bin("anchr")?;
+    let output = cmd
+        .arg("unitigs")
+        .arg("tests/Lambda/pe.cor.fa.gz")
+        .arg("tests/Lambda/env.json")
+        .arg("-o")
+        .arg("stdout")
+        .output()
+        .unwrap();
+    let stdout = String::from_utf8(output.stdout).unwrap();
+
+    assert!(stdout.lines().count() > 50);
+    assert!(stdout.contains("create_k_unitigs_large_k"));
+
+    let mut cmd = Command::cargo_bin("anchr")?;
+    let output = cmd
+        .arg("unitigs")
+        .arg("tests/Lambda/pe.cor.fa.gz")
+        .arg("tests/Lambda/env.json")
+        .arg("-u")
+        .arg("tadpole")
+        .arg("-o")
+        .arg("stdout")
+        .output()
+        .unwrap();
+    let stdout = String::from_utf8(output.stdout).unwrap();
+
+    assert!(stdout.lines().count() > 50);
+    assert!(!stdout.contains("create_k_unitigs_large_k"));
+    assert!(stdout.contains("tadpole"));
+
+    Ok(())
+}
+
+#[test]
 fn command_anchors() -> Result<(), Box<dyn std::error::Error>> {
     let mut cmd = Command::cargo_bin("anchr")?;
     let output = cmd